Includes update. Blueprint variable and blueprint linear combination moved to r1cs blueprint detail.
