+++ CMakeLists.txt
@@ -97,10 +97,9 @@ list(APPEND ${CURRENT_PROJECT_NAME}_PUBLIC_HEADERS
     include/nil/crypto3/zk/components/inner_product.hpp
     include/nil/crypto3/zk/components/loose_multiplexing.hpp
     include/nil/crypto3/zk/components/component_from_r1cs.hpp
-    include/nil/crypto3/zk/components/component.hpp
-    include/nil/crypto3/zk/components/blueprint_variable.hpp
-    include/nil/crypto3/zk/components/blueprint_linear_combination.hpp
-    include/nil/crypto3/zk/components/blueprint.hpp
+    include/nil/crypto3/zk/component.hpp
+    include/nil/crypto3/zk/blueprint/r1cs.hpp
+    include/nil/crypto3/zk/blueprint/plonk.hpp
     
     include/nil/crypto3/zk/merkle_tree.hpp)
