+++ include/nil/crypto3/zk/blueprint/r1cs.hpp
@@ -35,8 +35,8 @@
 
 #include <nil/crypto3/zk/snark/relations/constraint_satisfaction_problems/r1cs.hpp>
 
-#include <nil/crypto3/zk/blueprint_variable.hpp>
-#include <nil/crypto3/zk/blueprint_linear_combination.hpp>
+#include <nil/crypto3/zk/blueprint/detail/r1cs/blueprint_variable.hpp>
+#include <nil/crypto3/zk/blueprint/detail/r1cs/blueprint_linear_combination.hpp>
 
 namespace nil {
     namespace crypto3 {
