+++ include/nil/crypto3/zk/components/algebra/curves/edwards/r1cs/element_g1.hpp
@@ -34,8 +34,7 @@
 
 #include <nil/crypto3/zk/component.hpp>
 #include <nil/crypto3/zk/components/algebra/fields/element_fp.hpp>
-
-#include <nil/crypto3/zk/blueprint_variable.hpp>
+#include <nil/crypto3/zk/blueprint/r1cs.hpp>
 
 namespace nil {
     namespace crypto3 {
