+++ include/nil/crypto3/zk/components/algebra/curves/weierstrass/element_g2.hpp
@@ -31,15 +31,10 @@
 #ifndef CRYPTO3_ZK_BLUEPRINT_WEIERSTRASS_G2_COMPONENT_HPP
 #define CRYPTO3_ZK_BLUEPRINT_WEIERSTRASS_G2_COMPONENT_HPP
 
-#include <memory>
-
 #include <nil/crypto3/algebra/algorithms/pair.hpp>
 
 #include <nil/crypto3/zk/component.hpp>
-#include <nil/crypto3/zk/components/algebra/curves/detail/mnt4.hpp>
-#include <nil/crypto3/zk/components/algebra/curves/detail/mnt6.hpp>
-
-#include <nil/crypto3/zk/blueprint_variable.hpp>
+#include <nil/crypto3/zk/blueprint/r1cs.hpp>
 
 namespace nil {
     namespace crypto3 {
