+++ include/nil/crypto3/zk/components/algebra/fields/element_fp6_2over3.hpp
@@ -34,8 +34,7 @@
 #include <nil/crypto3/zk/component.hpp>
 #include <nil/crypto3/zk/components/algebra/fields/element_fp2.hpp>
 #include <nil/crypto3/zk/components/algebra/fields/element_fp3.hpp>
-
-#include <nil/crypto3/zk/blueprint_variable.hpp>
+#include <nil/crypto3/zk/blueprint/r1cs.hpp>
 
 namespace nil {
     namespace crypto3 {
