+++ include/nil/crypto3/zk/components/hashes/hash_io.hpp
@@ -31,8 +31,7 @@
 
 #include <nil/crypto3/zk/components/packing.hpp>
 #include <nil/crypto3/zk/component.hpp>
-#include <nil/crypto3/zk/components/blueprint.hpp>
-#include <nil/crypto3/zk/blueprint_variable.hpp>
+#include <nil/crypto3/zk/blueprint/r1cs.hpp>
 
 namespace nil {
     namespace crypto3 {
