+++ include/nil/crypto3/zk/components/hashes/poseidon/plonk/poseidon_9_wires.hpp
@@ -30,7 +30,6 @@
 
 #include <nil/crypto3/zk/blueprint/plonk.hpp>
 #include <nil/crypto3/zk/assignment/plonk.hpp>
-#include <nil/crypto3/zk/blueprint_variable.hpp>
 #include <nil/crypto3/zk/components/detail/plonk/n_wires.hpp>
 
 namespace nil {
