+++ include/nil/crypto3/zk/components/hashes/sha256/plonk/sha256.hpp
@@ -31,7 +31,6 @@
 #include <nil/crypto3/zk/components/packing.hpp>
 #include <nil/crypto3/zk/blueprint/plonk.hpp>
 #include <nil/crypto3/zk/assignment/plonk.hpp>
-#include <nil/crypto3/zk/blueprint_variable.hpp>
 
 namespace nil {
     namespace crypto3 {
