+++ include/nil/crypto3/zk/components/pubkey/eddsa/plonk/eddsa_5_wires.hpp
@@ -30,7 +30,6 @@
 
 #include <nil/crypto3/zk/blueprint/plonk.hpp>
 #include <nil/crypto3/zk/assignment/plonk.hpp>
-#include <nil/crypto3/zk/blueprint_variable.hpp>
 
 namespace nil {
     namespace crypto3 {
